Initial repository setup
{You are actually looking at the result of two history lifts. This
project started out in CVS. In January of 2009 it was moved to
Subversion.  On October 24 2011 it was moved to git, with the comment
history edited to turn Subversion commit references into references by
content and/or commit date, and massage comments into git form with a 
summary line. Here and elsewhere, comment portions in curly braces 
without preceding $ were added at conversion time.}