Revert some non-SCSI changes.
Oops! Dunno how the frick that happened folks! Sorry. This file
was NOT supposed to be part of the *SCSI* changes! (Duh!) And
I have no idea how the frick it got changed in the first place! But
in any case, the indentation of the #ifdef statements has been re-
stored back to the way it was. Sorry about that Bernard! - Fish