{POP conformance fix.}
Text in POP "The operation proceeds until a selected function
character equal to the test character is encountered, the 
second-operand location is exhausted, or a CPUdetermined
number of second-operand characters have been processed." 
r1 and r2 are both addresses and the CPU determined amount
of data is implemented when the last 12 bits is zero in r1 or
in r2. Fish take note of this before patching it back ;-)
Otherwise lets discuss and change the whole family to the
correct logic, but without warnings.