Type cast size_t to int as needed.
size_t is U64 on windows 64 bit machines. GCC defines size_t as
int. It is reasonable to type definitions as size_t when that are
relating to the sizeof() elements. For portability with windows
casting size_t to int should be alright since most of the left-sided
elements are usually int.