After reading the notes on the published ppt of the new machine..
I conclude that implementing nothing is ok. The enhancement facility, when
switched on, allows the cpu to work in blocks of data. This is what we
already do! Only the last block has a lot of calculations. I will
implement zero padding just like the real machine. We must keep
compatible! I wonder if the size of the padding will be published.