Fix on perfect fit writes within a page and the CMPSC enhancement facility.
Fix:
I implemented a turbo feature where main page addresses are stored. In
that way the absolute address does not need to be calculated every
symbol index or character. BUT when an index symbol or expanded index
symbol fitted exactly within the page, it also calculated the address of
the next page. This lead to an 0C4 (intc 4). 

This is fixed. In the case of a perfect match the stored address is
nullified and in the case of an adjecent needed page calculated.

CMPSC enhancement facility
The only thing I do (for now) is printing in debugging mode if zero
padding is requensted ;-) This is conform the POP.