Error in vstore.
When writing to multiple pages the buffer to write is devided into
chunks of 0x800 bytes. The Hercules page size. When there are more
chunks than 3 AND there is a perfect fit (last byte is on the last
byte of page 3), then cmpsc would faulty get the addres of the
adjecent page. This is fixed by testing >= 0x800 instead of > 0x800.