CMPSC changes:
1: I have received a debugging report from Bob where an expand is
started and finished right away. I suspect that initializing the MADDR
addresses for the dictionary raises an interrupt. Conform POP this
should not happen. I have changed CMPSC. It only fetches the dictionary
MADDR addresses on demand. It is faster also in this way ;-)
2: During debugging I also noticed that some OSs always have an
interrupt pending. In that way CMPSC returned with cc3 after reading or
writing 8 index symbols. Now the interrupt checking is done after a
minimum amount of uncompressed data processing. 32k to be precise.