CMPSC has been debugged together with Bob.
It has shown that the expansion stopped at the point where an
expansion entry was fetched.  hercules 3.05 worked fine. It showed
that the address was calculated with the MADDR macro, BUT the
ADDRESS_MAXWRAP was not used. It was in the old days. I hope that the
error is fixed now.

This svn is a debugging svn with debug messages. The next will be with
debugging off.