Report local modification in VERSION string.
Issue: using "3.07-svn-1234" as the VERSION string is great since it
tells us which SVN revision they're using (1234), but it doesn't tell
us whether they've made any local modifications to it or not
(i.e. whether they've made any of their own custom changes to
it). This patch addresses that issue.